added support for daily + workday and refactoring
A new function checkdate has been introduced to cleanup
the code for start, stop of EC2 and RDS.

'daily' and 'workkday' has beend added. This allow a much
smaller configuration in tags. It is needed for configurations
with multiple times for stoppping an instance, because a value
of a tag is limitted to 254 characters.
The data from daily or workday will be added to possible existing
data of the weekday.

'workday' is a placeholder for monday - friday.

Example:
{"daily": {"stop": [19,23,4]}, "sat": {"stop": 15},"sun": {"stop": [15]}}

A stop is planned for 4,19 and 23 for each day - including the addional
value for saturday and sunday.

This could be combined with "start" as well.